Home Shimmer, Detail Shimmer, SplashAnimation, DetailScreen, Scan Screen
