+++ app/src/main/java/com/bangkit/capstone/beangreader/presentation/screen/detail/component/FloatingFavorite.kt
@@ -8,6 +8,7 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import com.bangkit.capstone.beangreader.ui.theme.BeanGreaderTheme
 
@@ -23,7 +24,7 @@ fun FloatingFavorite(
         modifier = modifier
     ) {
         Icon(
-            tint = if (isFav) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.onSecondary,
+            tint = if (isFav) MaterialTheme.colorScheme.primary else Color.White,
             imageVector = if (isFav) Icons.Default.Favorite else Icons.Default.FavoriteBorder,
             contentDescription = "favorite_icon"
         )
