+++ app/src/main/java/com/bangkit/capstone/beangreader/presentation/screen/profile/ProfileScreen.kt
@@ -19,6 +19,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -85,6 +86,7 @@ fun ProfileContent(
             Text(
                 text = userData.username,
                 style = MaterialTheme.typography.titleLarge,
+                fontWeight = FontWeight.Medium,
                 maxLines = 1,
                 softWrap = true,
             )
