+++ app/src/main/java/com/bangkit/capstone/beangreader/presentation/screen/splash/SplashScreen.kt
@@ -1,21 +1,23 @@
 package com.bangkit.capstone.beangreader.presentation.screen.splash
 
+import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.animation.core.tween
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.bangkit.capstone.beangreader.R
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import kotlinx.coroutines.delay
-import kotlin.time.Duration.Companion.milliseconds
 
 @Composable
 fun SplashScreen(
@@ -25,28 +27,27 @@ fun SplashScreen(
 ) {
     val isLoggedIn by viewModel.isLoggedIn
     val currentOnTimeOut by rememberUpdatedState(onTimeOut)
-    val systemUiController = rememberSystemUiController()
 
-    DisposableEffect(key1 = true){
-        systemUiController.isSystemBarsVisible = false
-        onDispose {
-            systemUiController.isSystemBarsVisible = true
-        }
-    }
+    var startAnimation by remember { mutableStateOf(false) }
+    val alphaAnim = animateFloatAsState(
+        targetValue = if (startAnimation) 1f else 0f,
+        animationSpec = tween(2000)
+    )
     
     LaunchedEffect(currentOnTimeOut) {
-        delay(3000.milliseconds)
+        startAnimation = true
+        delay(3000L)
         onTimeOut(isLoggedIn)
     }
     
     Box(
         contentAlignment = Alignment.Center,
-        modifier = modifier
-            .fillMaxSize()
+        modifier = modifier.fillMaxSize()
     ) {
         Image(
             painter = painterResource(R.drawable.beanicon),
-            contentDescription = null
+            contentDescription = "logo",
+            alpha = alphaAnim.value
         )
     }
 }
